Refactoring & Cancel Stream Improvement
lint fixes
